+++ helper_functions.h
@@ -9,9 +9,9 @@
 #include <sstream>
 
 
-float *read_matrix_file(std::string filename, int N, int M);
+double*read_matrix_file(std::string filename, int N, int M);
 
-void print_matrix(float *matrix, int N, int M);
+void print_matrix(double*matrix, int N, int M);
 
 std::string *sequences_reader(std::string filename, int number_of_sequences);
